Add standard tests of ReadBehaviours and WritableBehaviours

+++ src/test/scala/com/cognite/sdk/scala/common/WritableBehaviors.scala
@@ -13,71 +13,87 @@ trait WritableBehaviors extends Matchers { this: FlatSpec =>
       idsThatDoNotExist: Seq[PrimitiveId],
       supportsMissingAndThrown: Boolean
   )(implicit t: Transformer[R, W]): Unit = {
-    it should "be an error to delete using ids that does not exist" in {
-      val thrown = the[CdpApiException] thrownBy writable
-        .deleteByIds(idsThatDoNotExist)
-      if (supportsMissingAndThrown) {
-        val missingIds = thrown.missing
-          .getOrElse(Seq.empty)
-          .flatMap(jsonObj => jsonObj("id").get.asNumber.get.toLong)
-        missingIds should have size idsThatDoNotExist.size.toLong
-        missingIds should contain theSameElementsAs idsThatDoNotExist
-      }
-
-      val sameIdsThatDoNotExist = Seq(idsThatDoNotExist.head, idsThatDoNotExist.head)
-      val sameIdsThrown = the[CdpApiException] thrownBy writable
-        .deleteByIds(sameIdsThatDoNotExist)
-      if (supportsMissingAndThrown) {
-        // as of 2019-06-03 we're inconsistent about our use of duplicated vs missing
-        // if duplicated ids that do not exist are specified.
-        val sameMissingIds = sameIdsThrown.duplicated match {
-          case Some(duplicatedIds) =>
-            duplicatedIds.flatMap(jsonObj => jsonObj("id").get.asNumber.get.toLong)
-          case None =>
-            sameIdsThrown.missing
-              .getOrElse(Seq.empty)
-              .flatMap(jsonObj => jsonObj("id").get.asNumber.get.toLong)
+    val datasetBaseUrl = "https://api.cognitedata.com/api/v1/projects/playground/datasets"
+    val deleteFlag =
+      if (!writable.baseUrl.toString().equals(datasetBaseUrl))
+        true
+      else
+        false
+    
+    if (deleteFlag) {
+      it should "be an error to delete using ids that does not exist" in {
+        val thrown = the[CdpApiException] thrownBy writable
+          .deleteByIds(idsThatDoNotExist)
+        if (supportsMissingAndThrown) {
+          val missingIds = thrown.missing
+            .getOrElse(Seq.empty)
+            .flatMap(jsonObj => jsonObj("id").get.asNumber.get.toLong)
+          (missingIds should have).size(idsThatDoNotExist.size.toLong)
+          (missingIds should contain).theSameElementsAs(idsThatDoNotExist)
         }
-        sameMissingIds should have size sameIdsThatDoNotExist.toSet.size.toLong
-        sameMissingIds should contain theSameElementsAs sameIdsThatDoNotExist.toSet
-      }
-    }
 
-    it should "include the request id in all cdp api exceptions" in {
-      val caught = intercept[CdpApiException](
-        writable.deleteByIds(idsThatDoNotExist)
-      )
-      assert(caught.requestId.isDefined)
+        val sameIdsThatDoNotExist = Seq(idsThatDoNotExist.head, idsThatDoNotExist.head)
+        val sameIdsThrown = the[CdpApiException] thrownBy writable
+          .deleteByIds(sameIdsThatDoNotExist)
+        if (supportsMissingAndThrown) {
+          // as of 2019-06-03 we're inconsistent about our use of duplicated vs missing
+          // if duplicated ids that do not exist are specified.
+          val sameMissingIds = sameIdsThrown.duplicated match {
+            case Some(duplicatedIds) =>
+              duplicatedIds.flatMap(jsonObj => jsonObj("id").get.asNumber.get.toLong)
+            case None =>
+              sameIdsThrown.missing
+                .getOrElse(Seq.empty)
+                .flatMap(jsonObj => jsonObj("id").get.asNumber.get.toLong)
+          }
+          (sameMissingIds should have).size(sameIdsThatDoNotExist.toSet.size.toLong)
+          (sameMissingIds should contain).theSameElementsAs(sameIdsThatDoNotExist.toSet)
+        }
+      }
+
+      it should "include the request id in all cdp api exceptions" in {
+        val caught = intercept[CdpApiException](
+          writable.deleteByIds(idsThatDoNotExist)
+        )
+        assert(caught.requestId.isDefined)
+      }
     }
 
     it should "create and delete items using the read class" in {
       // create a single item
       val createdItem = writable.createFromRead(readExamples.take(1))
-      createdItem should have size 1
+      (createdItem should have).size(1)
       createdItem.head.id should not be 0
-      writable.deleteByIds(createdItem.map(_.id))
+
+      if (deleteFlag)
+        writable.deleteByIds(createdItem.map(_.id))
 
       // create multiple items
       val createdItems = writable.createFromRead(readExamples)
-      createdItems should have size readExamples.size.toLong
+      (createdItems should have).size(readExamples.size.toLong)
       val createdIds = createdItems.map(_.id)
-      createdIds should have size readExamples.size.toLong
-      writable.deleteByIds(createdIds)
+      (createdIds should have).size(readExamples.size.toLong)
+
+      if (deleteFlag)
+        writable.deleteByIds(createdIds)
     }
 
     it should "create and delete items using the create class" in {
       // create a single item
       val createdItem = writable.create(createExamples.take(1))
-      createdItem should have size 1
+      (createdItem should have).size(1)
       createdItem.head.id should not be 0
-      writable.deleteByIds(createdItem.map(_.id))
+
+      if (deleteFlag)
+        writable.deleteByIds(createdItem.map(_.id))
 
       // create multiple items
       val createdItems = writable.create(createExamples)
-      createdItems should have size createExamples.size.toLong
+      (createdItems should have).size(createExamples.size.toLong)
       val createdIds = createdItems.map(_.id)
-      createdIds should have size createExamples.size.toLong
-      writable.deleteByIds(createdIds)
+      (createdIds should have).size(createExamples.size.toLong)
+      if (deleteFlag)
+        writable.deleteByIds(createdIds)
     }
     // TODO: test creating multiple items with the same external
     //       id in the same api call for V1
@@ -90,65 +106,95 @@ trait WritableBehaviors extends Matchers { this: FlatSpec =>
       externalIdsThatDoNotExist: Seq[String],
       supportsMissingAndThrown: Boolean
   )(implicit t: Transformer[R, W]): Unit = {
-    it should "be an error to delete using external ids that does not exist" in {
-      val thrown = the[CdpApiException] thrownBy writable
-        .deleteByExternalIds(externalIdsThatDoNotExist)
-      if (supportsMissingAndThrown) {
-        val missingIds = thrown.missing
-          .getOrElse(Seq.empty)
-          .flatMap(jsonObj => jsonObj("externalId").get.asString)
-        missingIds should have size externalIdsThatDoNotExist.size.toLong
-        missingIds should contain theSameElementsAs externalIdsThatDoNotExist
-      }
+    val datasetBaseUrl = "https://api.cognitedata.com/api/v1/projects/playground/datasets"
+    val deleteFlag =
+      if (!writable.baseUrl.toString().equals(datasetBaseUrl))
+        true
+      else
+        false
 
-      val sameIdsThatDoNotExist =
-        Seq(externalIdsThatDoNotExist.head, externalIdsThatDoNotExist.head)
-      val sameIdsThrown = the[CdpApiException] thrownBy writable
-        .deleteByExternalIds(sameIdsThatDoNotExist)
-      if (supportsMissingAndThrown) {
-        // as of 2019-06-03 we're inconsistent about our use of duplicated vs missing
-        // if duplicated ids that do not exist are specified.
-        val sameMissingIds = sameIdsThrown.duplicated match {
-          case Some(duplicatedIds) =>
-            duplicatedIds.flatMap(jsonObj => jsonObj("externalId").get.asString)
-          case None =>
-            sameIdsThrown.missing
-              .getOrElse(Seq.empty)
-              .flatMap(jsonObj => jsonObj("externalId").get.asString)
+    if (deleteFlag) {
+      it should "be an error to delete using external ids that does not exist" in {
+        val thrown = the[CdpApiException] thrownBy writable
+          .deleteByExternalIds(externalIdsThatDoNotExist)
+        if (supportsMissingAndThrown) {
+          val missingIds = thrown.missing
+            .getOrElse(Seq.empty)
+            .flatMap(jsonObj => jsonObj("externalId").get.asString)
+          (missingIds should have).size(externalIdsThatDoNotExist.size.toLong)
+          (missingIds should contain).theSameElementsAs(externalIdsThatDoNotExist)
         }
-        sameMissingIds should have size sameIdsThatDoNotExist.toSet.size.toLong
-        sameMissingIds should contain theSameElementsAs sameIdsThatDoNotExist.toSet
+
+        val sameIdsThatDoNotExist =
+          Seq(externalIdsThatDoNotExist.head, externalIdsThatDoNotExist.head)
+        val sameIdsThrown = the[CdpApiException] thrownBy writable
+          .deleteByExternalIds(sameIdsThatDoNotExist)
+        if (supportsMissingAndThrown) {
+          // as of 2019-06-03 we're inconsistent about our use of duplicated vs missing
+          // if duplicated ids that do not exist are specified.
+          val sameMissingIds = sameIdsThrown.duplicated match {
+            case Some(duplicatedIds) =>
+              duplicatedIds.flatMap(jsonObj => jsonObj("externalId").get.asString)
+            case None =>
+              sameIdsThrown.missing
+                .getOrElse(Seq.empty)
+                .flatMap(jsonObj => jsonObj("externalId").get.asString)
+          }
+          (sameMissingIds should have).size(sameIdsThatDoNotExist.toSet.size.toLong)
+          (sameMissingIds should contain).theSameElementsAs(sameIdsThatDoNotExist.toSet)
+        }
       }
     }
 
     it should "create and delete items using the read class and external ids" in {
       // create a single item
       val createdItem = writable.createFromRead(readExamples.take(1))
-      createdItem should have size 1
+      (createdItem should have).size(1)
       createdItem.head.externalId should not be empty
-      writable.deleteByExternalIds(createdItem.map(_.externalId.get))
+
+      if (deleteFlag)
+        writable.deleteByExternalIds(createdItem.map(_.externalId.get))
 
       // create multiple items
-      val createdItems = writable.createFromRead(readExamples)
-      createdItems should have size readExamples.size.toLong
-      val createdExternalIds = createdItems.map(_.externalId.get)
-      createdExternalIds should have size readExamples.size.toLong
-      writable.deleteByExternalIds(createdExternalIds)
+      if (deleteFlag) {
+        val createdItems = writable.createFromRead(readExamples)
+        (createdItems should have).size(readExamples.size.toLong)
+        val createdExternalIds = createdItems.map(_.externalId.get)
+        (createdExternalIds should have).size(readExamples.size.toLong)
+        writable.deleteByExternalIds(createdExternalIds)
+      }
+      else {
+        val createdItems = writable.createFromRead(Seq(readExamples(1),readExamples(2)))
+        (createdItems should have).size(2)
+        val createdExternalIds = createdItems.map(_.externalId.get)
+        (createdExternalIds should have).size(2)
+      }
     }
+
 
     it should "create and delete items using the create class and external ids" in {
       // create a single item
       val createdItem = writable.create(createExamples.take(1))
-      createdItem should have size 1
+      (createdItem should have).size(1)
       createdItem.head.externalId should not be empty
-      writable.deleteByExternalIds(createdItem.map(_.externalId.get))
+
+      if (deleteFlag)
+        writable.deleteByExternalIds(createdItem.map(_.externalId.get))
 
       // create multiple items
-      val createdItems = writable.create(createExamples)
-      createdItems should have size createExamples.size.toLong
-      val createdIds = createdItems.map(_.externalId.get)
-      createdIds should have size createExamples.size.toLong
-      writable.deleteByExternalIds(createdIds)
+      if (deleteFlag) {
+        val createdItems = writable.create(createExamples)
+        (createdItems should have).size(createExamples.size.toLong)
+        val createdIds = createdItems.map(_.externalId.get)
+        (createdIds should have).size(createExamples.size.toLong)
+        writable.deleteByExternalIds(createdIds)
+      }
+      else {
+        val createdItems = writable.create(Seq(createExamples(1),createExamples(2)))
+        (createdItems should have).size(2)
+        val createdIds = createdItems.map(_.externalId.get)
+        (createdIds should have).size(2)
+      }
     }
     // TODO: test creating multiple items with the same external
     //       id in the same api call for V1
@@ -166,6 +212,13 @@ trait WritableBehaviors extends Matchers { this: FlatSpec =>
       compareUpdated: (Seq[R], Seq[R]) => Unit
   )(implicit t: Transformer[R, W], t2: Transformer[R, U]): Unit =
     it should "allow updates using the read class" in {
+      val datasetBaseUrl = "https://api.cognitedata.com/api/v1/projects/playground/datasets"
+      val deleteFlag =
+        if (!updatable.baseUrl.toString().equals(datasetBaseUrl))
+          true
+        else
+          false
+
       // create items
       val createdIds = updatable.createFromRead(readExamples).map(_.id)
       assert(createdIds.size == readExamples.size)
@@ -189,7 +242,8 @@ trait WritableBehaviors extends Matchers { this: FlatSpec =>
       compareUpdated(readItems, updatedItems)
 
       // delete it
-      updatable.deleteByIds(createdIds)
+      if (deleteFlag)
+        updatable.deleteByIds(createdIds)
     }
 
   def updatableByExternalId[R <: WithExternalId, W, U](
@@ -202,10 +256,19 @@ trait WritableBehaviors extends Matchers { this: FlatSpec =>
     expectedBehaviors: (Seq[R], Seq[R]) => Unit)
   (implicit t: Transformer[R, W]): Unit =
     it should "allow updating by externalId" in {
+      val datasetBaseUrl = "https://api.cognitedata.com/api/v1/projects/playground/datasets"
+      val deleteFlag =
+        if (!resource.baseUrl.toString().equals(datasetBaseUrl))
+          true
+        else
+          false
+
       val createdItems = resource.createFromRead(itemsToCreate)
       val updatedItems = resource.updateByExternalId(updatesToMake)
       expectedBehaviors(createdItems, updatedItems)
-      resource.deleteByExternalIds(updatedItems.map(_.externalId.get))
+
+      if (deleteFlag)
+        resource.deleteByExternalIds(updatedItems.map(_.externalId.get))
     }
 
   def updatableById[R <: WithId[Long], W, U](
@@ -218,11 +281,20 @@ trait WritableBehaviors extends Matchers { this: FlatSpec =>
     expectedBehaviors: (Seq[R], Seq[R]) => Unit
   )(implicit t: Transformer[R, W]): Unit =
     it should "allow updating by Id" in {
+      val datasetBaseUrl = "https://api.cognitedata.com/api/v1/projects/playground/datasets"
+      val deleteFlag =
+        if (!resource.baseUrl.toString().equals(datasetBaseUrl))
+          true
+        else
+          false
+
       val createdItems = resource.createFromRead(itemsToCreate)
       val updateMap = createdItems.indices.map(i => (createdItems(i).id, updatesToMake(i))).toMap
       val updatedItems = resource.updateById(updateMap)
       expectedBehaviors(createdItems, updatedItems)
-      resource.deleteByIds(updatedItems.map(_.id))
+
+      if (deleteFlag)
+        resource.deleteByIds(updatedItems.map(_.id))
     }
 
   def deletableWithIgnoreUnknownIds[R <: WithId[Long], W, PrimitiveId](
